don't blindly jump to previous window
this causes a bug when the quickfix window doesn't open after grepping,
e.g. due to g:definitive_open_quickfix being set to 0. if it isn't open,
blindly doing wincmd p can cause the cursor to jump to a different
open window which is confusing. i will come up with a solution for this
at some point

... probably